Prepare 4.1.5 with Debian load-surface and warning cleanup

+++ src/DgDiscRFSGrids.h
@@ -302,9 +302,8 @@ template <class A, class B, class BG, class DB> class DgResAddConverter :
                         "from frame is not a DgDiscRFSGridsBase", DgBase::Fatal);
               }
 
-	      // JFW: second clause will never be > (int vs long):
               if (res() < 0 ||
-                  static_cast<unsigned long>(res()) >= discRFS().nRes() + 1)
+                  res() >= discRFS().nRes() + 1)
               {
                  report("DgResAddConverter<A, B, BG, DB>::DgResAddConverter() "
                         "invalid resolution", DgBase::Fatal);
@@ -365,9 +364,8 @@ template <class A, class B, class BG, class DB> class DgAddResConverter :
                         "from frame is not a DgDiscRFSGridsBase", DgBase::Fatal);
               }
 
-	      // JFW: second clause will never be > (int vs long):
               if (res() < 0 ||
-                  static_cast<unsigned long>(res()) >= discRFS().nRes() + 1)
+                  res() >= discRFS().nRes() + 1)
               {
                  report("DgAddResConverter<A, B, DB>::DgAddResConverter() "
                         "invalid resolution", DgBase::Fatal);
